+++ client/LandingPage.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react'
+
+export const LandingPage = () => (
+  <div>
+    <h2>Hello World</h2>
+    <a href="/calc">Click to open calculator</a>
+  </div>
+)
