+++ client/index.html
@@ -13,6 +13,7 @@
   </head>
   <body>
     <div id="root">
+      ${renderApp}
     </div>
     $bundles
   </body>
