--- config/be_config.js
@@ -1,12 +1,0 @@
-module.exports = function ServerConfig (isProduction) {
-  this.isProduction = isProduction
-  return {
-    target: "server@esnext",
-    autoImport: {
-      React: "react",
-      express: "express",
-      path: "path",
-      fs: "fs"
-    }
-  }
-}
