+++ config/be_config.ts
@@ -1,0 +1,6 @@
+export function ServerConfig(isProduction) {
+  this.isProduction = isProduction
+  return {
+    target: "server@esnext"
+  }
+}
