+++ package.json
@@ -4,9 +4,9 @@
   "description": "wt",
   "main": "index.js",
   "scripts": {
-    "start": "node fuse",
-    "be_serve": "node fuse be_serve",
-    "fe_serve": "node fuse fe_serve",
+    "start": "node -r ts-node/register --inspect fuse.ts",
+    "be_serve": "node fuse.ts be_serve",
+    "fe_serve": "node fuse.ts fe_serve",
     "be_test": "NODE_ENV=test mocha --compilers js:@babel/register --recursive 'server/test/'",
     "prod": "NODE_ENV=production run-s prebuild fe_build be_build be_run_build",
     "prebuild": "rm -rf build/*",
@@ -29,6 +29,10 @@
     "@babel/preset-env": "^7.1.0",
     "@babel/preset-react": "^7.0.0",
     "@babel/register": "^7.0.0",
+    "@types/express": "^4.16.0",
+    "@types/node": "^10.11.0",
+    "@types/react": "^16.4.14",
+    "@types/react-dom": "^16.0.7",
     "chai": "^4.1.2",
     "fuse-box": "^3.5.0",
     "mocha": "^5.2.0",
@@ -36,6 +40,7 @@
     "nodemon": "^1.18.4",
     "npm-run-all": "^4.1.3",
     "supertest": "^3.3.0",
+    "ts-node": "^7.0.1",
     "typescript": "^3.0.3",
     "uglify-es": "^3.3.9",
     "uglify-js": "^3.4.9"
